Collisions between paths
Today I worked on making a way so that if the player collides with an object while moving between paths, it is sent back to the path that it was already on. I Managed to find a way to do it with tags but it is very flawed. Ethan suggested another way which I am going to try next period.