Finished Level and other minor additions.
I finished creating the level that I sketched out and I also added a piece of UI which shows the speed of the player in a percentage from 0-100 and I added more comments to my code. I am nearing completion and need to add some instructions and more comments to my code and I should be finished.